Logit-based/requirements.txt
numpy>=1.26
pandas>=2.2
scikit-learn>=1.5
scipy>=1.11
pyyaml>=6.0
tqdm>=4.66
transformers>=4.43
torch>=2.3
openpyxl>=3.1
rich>=13.7
